requirements.txt
cycler==0.11.0
fonttools==4.34.4
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.1
seiche-aa-py @ git+https://github.com/Seiche-Ltd/Seiche-AA-py@c588b804cc0395d841c59c3d680f25d75fe5cbf9
six==1.16.0
